Solutions/ai-bartender-agent/aws/services/chat-api/src/chat-streaming/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
strands-agents>=1.0.0
bedrock-agentcore>=1.0.0
mcp-proxy-for-aws>=1.0.0
